project/requirements.txt
Django==4.2.5
django-rest-knox==4.2.0
djangorestframework==3.14.0
